requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.1
djangorestframework==3.16.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
